+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/tiles/BambooSpikesBlockTile.java
@@ -113,9 +113,9 @@ public class BambooSpikesBlockTile extends BlockEntity {
 
     public void makeParticle(Level level) {
         int i = this.getColor();
-        double d0 = (double) (i >> 16 & 255) / 255.0D;
-        double d1 = (double) (i >> 8 & 255) / 255.0D;
-        double d2 = (double) (i & 255) / 255.0D;
+        double d0 =  (i >> 16 & 255) / 255.0D;
+        double d1 = (i >> 8 & 255) / 255.0D;
+        double d2 = (i & 255) / 255.0D;
         BlockPos pos = this.getBlockPos();
         level.addParticle(ParticleTypes.ENTITY_EFFECT, pos.getX() + 0.5 + (level.random.nextFloat() - 0.5) * 0.75,
                 pos.getY() + 0.5 + (level.random.nextFloat() - 0.5) * 0.75,
